+++ include/hal/lra_hal.h
@@ -21,7 +21,7 @@
 
 void lra_hal_stop();
 bool lra_hal_init(uint8_t intensity);
-void lra_hal_task(uint32_t timestamp);
+void lra_hal_task(uint64_t timestamp);
 void lra_hal_push_amfm(haptic_processed_s *input);
 void lra_hal_set_standard(uint8_t intensity, bool brake);
 #endif
